+++ src/lib/hooks.ts
@@ -1,5 +1,6 @@
-import { useContext } from "react";
+import { useContext, useEffect, useState } from "react";
 import { feedbackItemsContext } from "../contexts/FeedbackItemsContextProvider";
+import { TFeedbackItem } from "./types";
 
 export function useFeedbackItemsContext() {
   const context = useContext(feedbackItemsContext);
@@ -9,4 +10,41 @@ export function useFeedbackItemsContext() {
     );
   }
   return context;
+}
+
+export function useFeedbackItems() {
+  const [feedbackItems, setFeedbackItems] = useState<TFeedbackItem[]>([]);
+  const [isLoading, setIsLoading] = useState(false);
+  const [errorMessage, setErrorMessage] = useState("");
+
+  useEffect(() => {
+    const fetchFeedbackItems = async () => {
+      setIsLoading(true);
+
+      try {
+        const response = await fetch(
+          "https://bytegrad.com/course-assets/projects/corpcomment/api/feedbacks"
+        );
+        if (!response.ok) {
+          throw new Error();
+        }
+        const data = await response.json();
+        setFeedbackItems(data.feedbacks);
+      } catch (error) {
+        setErrorMessage("Something went wrong. Please try again later");
+      }
+      setIsLoading(false);
+    };
+
+    fetchFeedbackItems();
+  }, []);
+
+  return {
+    feedbackItems,
+    isLoading,
+    errorMessage,
+    setFeedbackItems,
+    setIsLoading,
+    setErrorMessage,
+  };
 }
